move fetchMatchingRecyclers from MaterialFilter to RecyclerMainPage context
